Created make reservation (#43)
* Created make reservation method

+++ src/main/java/io/github/hotelmanagement/model/room/RoomRepository.java
@@ -4,9 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RoomRepository extends JpaRepository<Room, Integer> {
+public interface RoomRepository extends JpaRepository<Room, Long> {
     List<Room> findRoomsByBedAmount(int bedAmount);
-
     @Override
     <S extends Room> S save(S entity);
 }
